Add SVG images for Draw2D and Zest icons
This moves the SVG check from GEF to Draw2D so that it can be re-used,
also for Zest. Both bundles have been adapted to also use SVG icons if
supported by the active version of SWT.
