+++ org.eclipse.draw2d/META-INF/MANIFEST.MF
@@ -22,4 +22,5 @@ Require-Bundle: org.eclipse.swt;bundle-version="[3.126.0,4.0.0)";visibility:=ree
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Automatic-Module-Name: org.eclipse.draw2d
+Require-Capability: eclipse.swt;filter:="(image.format=svg)";resolution:=optional
