+++ org.eclipse.draw2d/src/org/eclipse/draw2d/CheckBox.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2024 IBM Corporation and others.
+ * Copyright (c) 2000, 2025 IBM Corporation and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -24,8 +24,8 @@ import org.eclipse.draw2d.internal.FileImageDataProvider;
 public final class CheckBox extends Toggle {
 	private Label label = null;
 
-	static final Image UNCHECKED = FileImageDataProvider.createImage(CheckBox.class, "images/checkboxenabledoff.png"); //$NON-NLS-1$
-	static final Image CHECKED = FileImageDataProvider.createImage(CheckBox.class, "images/checkboxenabledon.png"); //$NON-NLS-1$
+	static final Image UNCHECKED = FileImageDataProvider.createImage(CheckBox.class, "images/checkboxenabledoff.svg"); //$NON-NLS-1$
+	static final Image CHECKED = FileImageDataProvider.createImage(CheckBox.class, "images/checkboxenabledon.svg"); //$NON-NLS-1$
 
 	/**
 	 * Constructs a CheckBox with no text.
