+++ org.eclipse.gef/META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: %Plugin.name
 Bundle-SymbolicName: org.eclipse.gef; singleton:=true
-Bundle-Version: 3.22.100.qualifier
+Bundle-Version: 3.23.0.qualifier
 Bundle-Activator: org.eclipse.gef.internal.InternalGEFPlugin
 Bundle-Vendor: %Plugin.providerName
 Bundle-Localization: plugin
@@ -31,7 +31,7 @@ Export-Package: org.eclipse.gef,
  org.eclipse.gef.ui.rulers,
  org.eclipse.gef.ui.views.palette,
  org.eclipse.gef.util
-Require-Bundle: org.eclipse.draw2d;visibility:=reexport;bundle-version="[3.13.0,4.0.0)",
+Require-Bundle: org.eclipse.draw2d;visibility:=reexport;bundle-version="[3.20.100,4.0.0)",
  org.eclipse.core.runtime;bundle-version="[3.5.0,4.0.0)",
  org.eclipse.ui.views;resolution:=optional;bundle-version="[3.2.0,4.0.0)",
  org.eclipse.ui.workbench;bundle-version="[3.133.0,4.0.0)",
