+++ org.eclipse.zest.core/META-INF/MANIFEST.MF
@@ -4,10 +4,10 @@ Bundle-Name: %Plugin.name
 Bundle-SymbolicName: org.eclipse.zest.core;singleton:=true
 Bundle-Vendor: %Plugin.providerName
 Bundle-Localization: plugin
-Bundle-Version: 1.15.200.qualifier
+Bundle-Version: 1.16.0.qualifier
 Require-Bundle: org.eclipse.zest.layouts,
  org.eclipse.ui;bundle-version="[3.2.0,4.0.0)",
- org.eclipse.draw2d;visibility:=reexport
+ org.eclipse.draw2d;bundle-version="[3.20.100,4.0.0)";visibility:=reexport
 Bundle-ActivationPolicy: lazy
 Export-Package: org.eclipse.zest.core.viewers,
  org.eclipse.zest.core.viewers.internal;x-internal:=true,
@@ -17,3 +17,4 @@ Export-Package: org.eclipse.zest.core.viewers,
  org.eclipse.zest.core.widgets.internal;x-friends:="org.eclipse.zest.tests"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Automatic-Module-Name: org.eclipse.zest.core
+Require-Capability: eclipse.swt;filter:="(image.format=svg)";resolution:=optional
